allusion mostly done
