+++ src/app/api/allusions/allusionQuery/route.js
@@ -1,68 +1,145 @@
 const { getSession } = require('../../neo4j_driver/route.js');
-const { toNativeTypes, valueToNativeType, getChpList,concatObj } = require('../../neo4j_driver/utils.js');
+const { toNativeTypes, valueToNativeType,concatObj, sortPnumsFromObjList } = require('../../neo4j_driver/utils.js');
 
 
 async function getData (){
 	const session = await getSession();
 
-	//all the get method and return the db data
-	const queries = {
-		getHonka:  'match (a:Honka) return (a) as honka',
-		getPnum: 'match (g:Genji_Poem) return g.pnum as pnum',
-		getPoemHonka: 'MATCH (n:Honka)-[r:ALLUDES_TO]-(p:Genji_Poem) RETURN n.id as id, p.pnum as pnum, r.notes as notes',
-		getPoet: 'match (p:People) return p.name as poet',
-		getHonkaPoet: 'match (h:Honka)<-[:AUTHOR_OF]-(p:People) return h.id as id, p.name as name',
-		getSource: 'match (s:Source) return s.title as source',
-		getHonkaSource: 'match (h:Honka)-[r:ANTHOLOGIZED_IN]-(s:Source) return h.id as id, r.order as order, s.title as title',
-		getTrans: 'match (h:Honka)<-[:TRANSLATION_OF]-(t:Translation)<-[:TRANSLATOR_OF]-(p:People) return h.id as id, t.translation as trans, p.name as name'
-	};
+	// All the Quearies that return the DB data
+	let getHonka = 'match (a:Honka) return (a) as honka'
+	let getPnum = 'match (g:Genji_Poem) return g.pnum as pnum'
+	let getPoemHonka = 'MATCH (n:Honka)-[r:ALLUDES_TO]-(p:Genji_Poem) RETURN n.id as id, p.pnum as pnum, r.notes as notes'
+	let getPoet = 'match (p:People) return p.name as poet'
+	let getHonkaPoet = 'match (h:Honka)<-[:AUTHOR_OF]-(p:People) return h.id as id, p.name as name'
+	let getSource = 'match (s:Source) return s.title as source'
+	let getHonkaSource = 'match (h:Honka)-[r:ANTHOLOGIZED_IN]-(s:Source) return h.id as id, r.order as order, s.title as title'
+	let getTrans = 'match (h:Honka)<-[:TRANSLATION_OF]-(t:Translation)<-[:TRANSLATOR_OF]-(p:People) return h.id as id, t.translation as trans, p.name as name'
 
 
-
-	const result = {};
 	try {
-		for (let key in queries) {
-			const queryResult = await session.readTransaction(tx => 
-				tx.run(queries[key])
-			); 
-			result[key] = queryResult;
-		} 
+		const resHonka = await session.readTransaction(tx => tx.run(getHonka))
+		const resPnum = await session.readTransaction(tx => tx.run(getPnum))
+		const resPoemHonka = await session.readTransaction(tx => tx.run(getPoemHonka))
+		const resPoet = await session.readTransaction(tx => tx.run(getPoet))
+		const resPoetEdge = await session.readTransaction(tx => tx.run(getHonkaPoet))
+		const resSrc = await session.readTransaction(tx => tx.run(getSource))
+		const resSourceEdge = await session.readTransaction(tx => tx.run(getHonkaSource))
+		const resTrans = await session.readTransaction(tx => tx.run(getTrans))
 
-		let exchange = new Set()           
-		result['res'].records.map(e => JSON.stringify(toNativeTypes(e.get('exchange')))).forEach(e => exchange.add(e))
-		exchange = Array.from(exchange).map(e => JSON.parse(e))
-		
-		//for transtemp
-		let transTemp = result['res'].records.map(e => toNativeTypes(e.get('trans'))).map(e => [e.end.properties.name, e.segments[0].end.properties.translation, e.segments[1].start.properties.WaleyPageNum])
-		
-		let sources = result['resHonkaInfo'].records.map(e => [Object.values(toNativeTypes(e.get('honka'))).join(''), Object.values(toNativeTypes(e.get('title'))).join(''), Object.values(toNativeTypes(e.get('romaji'))).join(''), Object.values(toNativeTypes(e.get('poet'))).join(''), Object.values(toNativeTypes(e.get('order'))).join(''), Object.values(toNativeTypes(e.get('translator'))).join(''), Object.values(toNativeTypes(e.get('translation'))).join(''), e.get('notes') !== null ? Object.values(toNativeTypes(e.get('notes'))).join('') : 'N/A'])
-		
-		//related
-		let related = new Set()
-		result['resRel'].records.map(e => toNativeTypes(e.get('rel'))).forEach(e => related.add([Object.values(e).join('')]))
-		related = Array.from(related).flat()
-		related = related.map(e => [e, true])
-		//res tag
-		let tags = new Set()
-		result['resTag'].records.map(e => toNativeTypes(e.get('type'))).forEach(e => tags.add([Object.values(e).join('')]))
-		tags = Array.from(tags).flat()
-		tags = tags.map(e => [e, true])
-		//types
-		let types = result['resType'].records.map(e => e.get('type'))
-		//ls
-		let ls = []
-		types.forEach(e => ls.push({value: e, label: e})) 
-		//pls
-		let temp = result['resPnum'].records.map(e => e.get('pnum'))
-		let pls = []
-		temp.forEach(e => {
-			pls.push({value:e, label:e})
+		// Return variables
+		let ans = []
+		let max = 0
+		let key = 0
+		let translators = new Set()
+
+
+		// For translators
+    	let transLs = resTrans.records.map(e => [concatObj(toNativeTypes(e.get('id'))), concatObj(toNativeTypes(e.get('trans'))), concatObj(toNativeTypes(e.get('name')))])
+		let transObj = {}
+		transLs.forEach(e => {
+			translators.add(e[2])
+			if (transObj[e[0]] === undefined) {
+				transObj[e[0]] = {}
+				transObj[e[0]][e[2]] = e[1]
+			} else {
+				transObj[e[0]][e[2]] = e[1]
+			}
 		})
-		const data = [exchange, transTemp, sources, related, tags, ls, pls];
+		translators = Array.from(translators).map(e => ({value: e, label: e}))
+
+
+		// For ans
+		resHonka.records.map(e => toNativeTypes(e.get('honka'))).forEach(e => {
+			delete Object.assign(e.properties, { ['key']: e.properties['id'] })['id']
+			e.properties.translations = transObj[e.properties.key]
+			ans.push(e.properties)
+			key = parseInt(e.properties.key.slice(1))
+			if (max < key) {
+				max = key
+			}
+		})
+		let tempEdgeList = resSourceEdge.records.map(e => [concatObj(toNativeTypes(e.get('id'))), concatObj(toNativeTypes(e.get('title'))), concatObj(toNativeTypes(e.get('order')))])
+		let tempEdgeObj = {}
+		tempEdgeList.forEach(e => {
+			if (tempEdgeObj[e[0]] === undefined) {
+				tempEdgeObj[e[0]] = [[e[1], e[2], true]]
+			} else {
+				tempEdgeObj[e[0]].push([e[1], e[2], true])
+			}
+		})
+		ans.forEach(e => {
+			if (tempEdgeObj[e.key] !== undefined) {
+				e.Source = tempEdgeObj[e.key]
+			} 
+		})
+		let poetEdge = resPoetEdge.records.map(e => [concatObj(toNativeTypes(e.get('id'))), concatObj(toNativeTypes(e.get('name')))])
+		poetEdge.forEach(e => {
+			let index = ans.findIndex(ele => ele.key === e[0])
+			ans[index].Poet = e[1]
+		})
+
+
+		// For init_src and init_order
+		let init_src = {}
+		let init_order = {}
+		for (let i = 0; i < max; i ++) {
+			init_src['H'+JSON.stringify(i)] = ''
+			init_order['H'+JSON.stringify(i)] = 'N/A'
+		}
+
+
+		// For ls
+		let temp = resPnum.records.map(e => e.get('pnum'))
+		let ls = []
+		temp.forEach(e => {
+			if (e !== null) {
+				ls.push({ value: e, label: e })
+			}
+		})
+		ls = sortPnumsFromObjList(ls)
+
+
+		// For links
+		let ll = Array.from(new Set(resPoemHonka.records.map(e => JSON.stringify([e.get('id'), e.get('pnum'), e.get('notes')])))).map(e => JSON.parse(e))
+		let links = {}
+		ll.forEach(e => {
+			if (e[0] in links) {
+				links[e[0]].push([e[1], true, e[2]])
+			} else {
+				links[e[0]] = [[e[1], true, e[2]]]
+			}
+		})
+	
+
+		// For poets
+		temp = resPoet.records.map(e => e.get('poet'))
+		let poets = []
+		temp.forEach(e => {
+			poets.push({ value: e, label: e })
+		})
+
+		// For sources
+		temp = resSrc.records.map(e => e.get('source'))
+		let sources = []
+		temp.forEach(e => {
+			sources.push({ value: e, label: e })
+		})
+		
+		
+		const data = {
+			translators: translators, 
+			ans: ans, 
+			init_src: init_src,
+			init_order: init_order,
+			ls:ls,
+			links:links,
+			poets:poets,
+			sources:sources,
+			max: max
+		};
 		return (data);
 	} catch(error) {
-		console.error('Failed to execute queries in poems:', error);
-		result.status(500).json({ error: 'Failed to execute queries' });
+		console.error('Failed to execute queries in allusions/allusionQuery:', error);
 	} finally{
 		await session.close();
 	}
@@ -70,7 +147,7 @@ async function getData (){
 
 export const GET = async (request) =>{
 	try {   
-		const data = await getData(chapter, number)
+		const data = await getData()
 		return new Response(JSON.stringify(data), {status: 200})
 	}catch (error){
 		return new Response(error, {status: 500})
