+++ src/app/api/neo4j_driver/utils.js
@@ -245,4 +245,4 @@ export function generateGeneology(l) {
   }
   
 
-module.exports = { toNativeTypes, valueToNativeType, getChpList, generateGeneology, concatObj, getPoemTableContent }
+module.exports = { toNativeTypes, valueToNativeType, getChpList, generateGeneology, concatObj, getPoemTableContent, sortPnumsFromObjList }
